requirements.txt
langchain
openai
pandasai
streamlit
streamlit-chat
tiktoken
faiss-cpu
python-dotenv
chromadb
